Add algorithm.h and portability/fixes
Add new opencstl/algorithm.h and include it in the umbrella header; update amalgamate.py to copy the single-header into examples. Fix various portability and correctness issues: change __cstl_va_start to cast beg=(void*)V, widen RNG state to unsigned __int128 and cast return to unsigned long long, improve platform-specific time functions and fopen behavior, and rename version symbols to OPENCSTL_VERSION/opencstl_version. Tidy up main and tests (limit rand ranges, fix printf formats, reorder test calls, remove dead code) and clean test/check.c. Update README platform table and formatting and make a small CMakeLists tweak.

+++ opencstl/cstl_random.h
@@ -40,8 +40,11 @@
 #include <limits.h>
 #define CSTL_RAND64_MAX ULLONG_MAX
 #define CSTL_RAND32_MAX UINT_MAX
-static unsigned long long _seed64 = 1;
-static unsigned long _seed32 = 1;
+static unsigned __int128 _seed64 = 1;
+static unsigned long long _seed32 = 1;
+
+//static unsigned long long _seed64 = 1;
+//static unsigned int _seed32 = 1;
 
 static void cstl_rand_seed32(unsigned int seed) {
     _seed32 = seed;
@@ -58,7 +61,7 @@ static void cstl_rand_seed64(unsigned long long seed) {
 
 static unsigned long long cstl_rand64() {
     _seed64 = (_seed64 * 6364136223846793005ULL) + 1442695040888963407ULL;
-    return _seed64;
+    return (unsigned long long) _seed64;
 }
